+++ pamqp/specification.py
@@ -108,11 +108,10 @@ class Frame(object):
         """Return an attribute as if it were a dict.
 
         :param str item: The item to look for
+        :raises: KeyError
         :rtype: any
 
         """
-        if item not in self.__slots__:
-            return None
         return getattr(self, item)
 
     def __len__(self):
@@ -131,6 +130,15 @@ class Frame(object):
         """
         return '<%s.%s object at %s>' % (__name__, self.name, hex(id(self)))
 
+    @classmethod
+    def type(cls, attr):
+        """Return the data type for an attribute.
+
+        :rtype: str
+
+        """
+        return getattr(cls, '_' + attr)
+
     def marshal(self):
         """
         Dynamically encode the frame by taking the list of attributes and
@@ -145,7 +153,7 @@ class Frame(object):
         byte = None
         offset = 0
         for argument in self.__slots__:
-            data_type = getattr(self.__class__, '_' + argument)
+            data_type = self.type(argument)
 
             # Check if we need to turn on bit processing
             if not processing_bitset and data_type == 'bit':
@@ -199,8 +207,7 @@ class Frame(object):
         offset = 0
         processing_bitset = False
         for argument in self.__slots__:
-
-            data_type = getattr(self.__class__, '_' + argument)
+            data_type = self.type(argument)
 
             if offset == 7 and processing_bitset:
                 data = data[1:]
@@ -234,15 +241,37 @@ class PropertiesBase(object):
     name = 'PropertiesBase'
 
     def __contains__(self, item):
-        return item in self.__slots__ and getattr(self, item, None)
-
-    def __iter__(self):
-        for attribute in self.__slots__:
-            yield attribute
+        return item in self.__slots__
 
     def __delattr__(self, item):
-        if item in self.__slots__:
-            setattr(self, item, None)
+        setattr(self, item, None)
+
+    def __iter__(self):
+        """Iterate the attributes and values as key, value pairs.
+
+        :rtype: tuple
+
+        """
+        for attribute in self.__slots__:
+            yield (attribute, getattr(self, attribute))
+
+    @classmethod
+    def attributes(cls):
+        """Return the list of attributes
+
+        :rtype: list
+
+        """
+        return [attr for attr in cls.__slots__]
+
+    @classmethod
+    def type(cls, attr):
+        """Return the data type for an attribute.
+
+        :rtype: str
+
+        """
+        return getattr(cls, '_' + attr)
 
     def encode_property(self, property_name, property_value):
         """Encode a single property value
@@ -251,8 +280,7 @@ class PropertiesBase(object):
         :param any property_value: The value to encode
 
         """
-        return encode.by_type(property_value,
-                              getattr(self.__class__, '_' + property_name))
+        return encode.by_type(property_value, self.type(property_name))
 
     def marshal(self):
         """Take the Basic.Properties data structure and marshal it into the data
@@ -282,10 +310,12 @@ class PropertiesBase(object):
         return b''.join(flag_pieces + parts)
 
     def to_dict(self):
-        output = dict()
-        for attribute in self.__slots__:
-            output[attribute] = getattr(self, attribute, None)
-        return output
+        """Return the properties as a dict
+
+        :rtype: dict
+
+        """
+        return dict(self)
 
     def unmarshal(self, flags, data):
         """
@@ -518,6 +548,8 @@ class Connection(object):
     thereafter.
 
     """
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 10
     index = 0x000A0000
@@ -930,9 +962,6 @@ class Connection(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
     class Blocked(Frame):
         """Signal that connection is blocked
 
@@ -977,9 +1006,6 @@ class Connection(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
 
 class Channel(object):
     """Work with channels
@@ -988,6 +1014,8 @@ class Channel(object):
     server and for both peers to operate the channel thereafter.
 
     """
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 20
     index = 0x00140000
@@ -1189,9 +1217,6 @@ class Channel(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
 
 class Exchange(object):
     """Work with exchanges
@@ -1200,6 +1225,8 @@ class Exchange(object):
     configured in the server or declared at runtime.
 
     """
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 40
     index = 0x00280000
@@ -1305,9 +1332,6 @@ class Exchange(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
     class Delete(Frame):
         """Delete an exchange
 
@@ -1373,9 +1397,6 @@ class Exchange(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
     class Bind(Frame):
         # AMQP Method Number and Mapping Index
         frame_id = 30
@@ -1438,9 +1459,6 @@ class Exchange(object):
 
         # Specifies if this is a synchronous AMQP method
         synchronous = False
-
-        # AMQP Method Attributes
-        attributes = []
 
     class Unbind(Frame):
         # AMQP Method Number and Mapping Index
@@ -1505,9 +1523,6 @@ class Exchange(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
 
 class Queue(object):
     """Work with queues
@@ -1517,6 +1532,8 @@ class Queue(object):
     order to receive messages from publishers.
 
     """
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 50
     index = 0x00320000
@@ -1717,9 +1734,6 @@ class Queue(object):
 
         # Specifies if this is a synchronous AMQP method
         synchronous = False
-
-        # AMQP Method Attributes
-        attributes = []
 
     class Purge(Frame):
         """Purge a queue
@@ -1948,9 +1962,6 @@ class Queue(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
 
 class Basic(object):
     """Work with basic content
@@ -1959,6 +1970,8 @@ class Basic(object):
     messaging model.
 
     """
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 60
     index = 0x003C0000
@@ -2027,9 +2040,6 @@ class Basic(object):
 
         # Specifies if this is a synchronous AMQP method
         synchronous = False
-
-        # AMQP Method Attributes
-        attributes = []
 
     class Consume(Frame):
         """Start a queue consumer
@@ -2657,9 +2667,6 @@ class Basic(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
     class Nack(Frame):
         # AMQP Method Number and Mapping Index
         frame_id = 120
@@ -2699,21 +2706,20 @@ class Basic(object):
 
         name = 'Basic.Properties'
 
-        # Attributes
         __slots__ = ['content_type',
-                      "content_encoding",
-                      "headers",
-                      "delivery_mode",
-                      "priority",
-                      "correlation_id",
-                      "reply_to",
-                      "expiration",
-                      "message_id",
-                      "timestamp",
-                      "message_type",
-                      "user_id",
-                      "app_id",
-                      "cluster_id"]
+                     'content_encoding',
+                     'headers',
+                     'delivery_mode',
+                     'priority',
+                     'correlation_id',
+                     'reply_to',
+                     'expiration',
+                     'message_id',
+                     'timestamp',
+                     'message_type',
+                     'user_id',
+                     'app_id',
+                     'cluster_id']
 
         # Flag Values
         flags = {'content_type': 32768,
@@ -2832,6 +2838,8 @@ class Tx(object):
     mandatory flags on Basic.Publish methods is not defined.
 
     """
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 90
     index = 0x005A0000
@@ -2855,9 +2863,6 @@ class Tx(object):
         # Valid responses to this method
         valid_responses = ['Tx.SelectOk']
 
-        # AMQP Method Attributes
-        attributes = []
-
     class SelectOk(Frame):
         """Confirm transaction mode
 
@@ -2872,9 +2877,6 @@ class Tx(object):
 
         # Specifies if this is a synchronous AMQP method
         synchronous = False
-
-        # AMQP Method Attributes
-        attributes = []
 
     class Commit(Frame):
         """Commit the current transaction
@@ -2895,9 +2897,6 @@ class Tx(object):
         # Valid responses to this method
         valid_responses = ['Tx.CommitOk']
 
-        # AMQP Method Attributes
-        attributes = []
-
     class CommitOk(Frame):
         """Confirm a successful commit
 
@@ -2912,9 +2911,6 @@ class Tx(object):
 
         # Specifies if this is a synchronous AMQP method
         synchronous = False
-
-        # AMQP Method Attributes
-        attributes = []
 
     class Rollback(Frame):
         """Abandon the current transaction
@@ -2937,9 +2933,6 @@ class Tx(object):
         # Valid responses to this method
         valid_responses = ['Tx.RollbackOk']
 
-        # AMQP Method Attributes
-        attributes = []
-
     class RollbackOk(Frame):
         """Confirm successful rollback
 
@@ -2955,11 +2948,10 @@ class Tx(object):
         # Specifies if this is a synchronous AMQP method
         synchronous = False
 
-        # AMQP Method Attributes
-        attributes = []
-
 
 class Confirm(object):
+    __slots__ = []
+
     # AMQP Class Number and Mapping Index
     frame_id = 85
     index = 0x00550000
@@ -2999,9 +2991,6 @@ class Confirm(object):
 
         # Specifies if this is a synchronous AMQP method
         synchronous = False
-
-        # AMQP Method Attributes
-        attributes = []
 
 # AMQP Class.Method Index Mapping
 INDEX_MAPPING = {0x000A000A: Connection.Start,